server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0
python-dotenv==1.0.0